打可部署jar包 框架升级

+++ ev-gb-gateway-tcu/src/main/java/com/dyy/tsp/evgb/gateway/tcu/enumtype/TcuCoreType.java
@@ -69,7 +69,7 @@ public enum TcuCoreType {
 
     public static TcuCoreType valuesOf(Short id) {
         for (TcuCoreType enums : TcuCoreType.values()) {
-            if (enums.getId()==id) {
+            if (enums.getId()==id.shortValue()) {
                 return enums;
             }
         }
